Got rid of signed/unsigned compiler warnings
git-svn-id: http://svn.code.sf.net/p/coda-oss/code/trunk@719 1dbd42e5-0d20-4692-8756-183d5fc073b5

+++ modules/c++/cli/include/cli/Value.h
@@ -122,7 +122,7 @@ public:
     /**
      * Returns the size of the value.
      */
-    unsigned int size() const
+    size_t size() const
     {
         return mValues.size();
     }
